// frontend/src/api/schema.ts
export interface components {
    schemas: {
// ...
        DateRangeFilter: "today" | "week" | "month";
// ...
        ReportShortResponse: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            status: components["schemas"]["ReportStatus"];
            /**
             * Reporter Id
             * Format: uuid
             */
            reporter_id: string;
            /** Drug Name */
            drug_name?: string | null;
            /** Adverse Reaction */
            adverse_reaction?: string | null;
            /** Reporter Name */
            reporter_name?: string | null;
            /** Is Clinically Significant */
            is_clinically_significant?: boolean | null;
            /** Analysis Status */
            analysis_status: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            severity?: components["schemas"]["SeverityLevel"] | null;
        };
        /**
         * ReportStatus
         * @enum {string}
         */
        ReportStatus: "submitted" | "clarification" | "analysis" | "finalized";
        /**
         * SeverityLevel
         * @enum {string}
         */
        SeverityLevel: "mild" | "moderate" | "severe" | "life-threatening";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}